move CommonJsChunkFormat into separate plugin
